Move payment screen files into their own package
